requirements.txt
# NEURON — Dependencies

# Data collection
numpy>=1.24             # Array operations for snapshot stats
python-osc>=1.8         # OSC protocol (Mind Monitor streaming)
pylsl>=1.16             # Lab Streaming Layer (Petal Metrics / muselsl)

# Configuration
pyyaml>=6.0
python-dotenv>=1.0      # .env file support

# Terminal UI (CLI mode)
rich>=13.0

# Web server
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.9

# Claude API
anthropic>=0.52
claude-agent-sdk>=0.1

# AI image generation (Nano Banana 2)
google-genai>=1.0

# AI music generation (ElevenLabs)
elevenlabs>=1.0

# No scipy, no mne, no sklearn — Claude does the interpretation.
